requirements.txt
google-genai
google-adk
pandas
numpy
scikit-learn
statsmodels
jinja2
python-dotenv
pytest
pytest-asyncio
